requirements.txt
zhipuai>=2.0.0
pyyaml>=6.0
jinja2>=3.1.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=4.9.0